Refactor to kmp and restructure demo application
