+++ demo/src/main/kotlin/com/kdroid/composetray/demo/App.kt
@@ -1,4 +1,4 @@
-package sample
+package com.kdroid.composetray.demo
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.layout.*
