+++ demo/src/main/kotlin/com/kdroid/composetray/demo/DemoWithContextMenu.kt
@@ -1,8 +1,9 @@
-package sample
-import androidx.compose.foundation.Image
+package com.kdroid.composetray.demo
+
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Icon
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Notifications
+import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -10,14 +11,11 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
-import androidx.compose.ui.window.rememberWindowState
 import com.kdroid.composetray.tray.api.Tray
 import com.kdroid.composetray.utils.SingleInstanceManager
 import com.kdroid.composetray.utils.getTrayPosition
-import com.kdroid.composetray.utils.getTrayWindowPosition
 import com.kdroid.kmplog.Log
 import com.kdroid.kmplog.d
 import com.kdroid.kmplog.i
@@ -42,18 +40,17 @@ fun main() = application {
         return@application
     }
 
-
     // Updated condition for Tray visibility
     if (alwaysShowTray || !isWindowVisible) {
         Tray(
-           iconContent = {
-               Image(
-                   Icons.Default.Notifications,
-                   contentDescription = "Application Icon",
-                   modifier = Modifier.fillMaxSize(),
-                   colorFilter = androidx.compose.ui.graphics.ColorFilter.tint(Color.White)
-               )
-           },
+            iconContent = {
+                Icon(
+                    Icons.Default.Favorite,
+                    contentDescription = "",
+                    tint = Color.White,
+                    modifier = Modifier.fillMaxSize()
+                )
+            },
             primaryAction = {
                 isWindowVisible = true
                 Log.i(logTag, "On Primary Clicked")
@@ -61,6 +58,27 @@ fun main() = application {
             primaryActionLinuxLabel = "Open the Application",
             tooltip = "My Application"
         ) {
+            // Options SubMenu
+            SubMenu(label = "Options") {
+                Item(label = "Show Text") {
+                    Log.i(logTag, "Show Text selected")
+                    textVisible = true
+                }
+                Item(label = "Hide Text") {
+                    Log.i(logTag, "Hide Text selected")
+                    textVisible = false
+                }
+                SubMenu(label = "Advanced Sub-options") {
+                    Item(label = "Advanced Option 1") {
+                        Log.i(logTag, "Advanced Option 1 selected")
+                    }
+                    Item(label = "Advanced Option 2") {
+                        Log.i(logTag, "Advanced Option 2 selected")
+                    }
+                }
+            }
+
+            Divider()
 
             // Tools SubMenu
             SubMenu(label = "Tools") {
@@ -102,6 +120,9 @@ fun main() = application {
 
             Divider()
 
+            Item(label = "Hide in tray") {
+                isWindowVisible = false
+            }
 
             Item(label = "Exit", isEnabled = true) {
                 Log.i(logTag, "Exiting the application")
@@ -113,11 +134,6 @@ fun main() = application {
         }
     }
 
-
-    val windowWidth = 800
-    val windowHeight = 600
-    val windowPosition = getTrayWindowPosition(windowWidth, windowHeight)
-
     Window(
         onCloseRequest = {
             if (hideOnClose) {
@@ -126,11 +142,6 @@ fun main() = application {
                 exitApplication()
             }
         },
-        state = rememberWindowState(
-            width = windowWidth.dp,
-            height = windowHeight.dp,
-            position = windowPosition
-        ),
         title = "Compose Desktop Application with Two Screens",
         visible = isWindowVisible,
         icon = painterResource("icon.png") // Optional: Set window icon
