+++ demo/src/main/resources/icon.ico
