+++ src/commonMain/kotlin/com/kdroid/composetray/lib/windows/StdCallCallback.kt
