+++ src/commonMain/kotlin/com/kdroid/composetray/utils/SingleInstanceManager.kt
